"Implement" internal modules as a single import. Also establish somewhat of a "uniform design" for all of them. Let's see how long it will take me to discover arcane magicks known as "interfaces".
Still think it's largely garbage but at least the final result and functionality check out with what I initially wanted.

Switch TS from 3.7 to 3.8 because I'm a retard and couldn't figure out a clean way of exposing internal modules to the main thing.
Blessed be "export * as <name>", for saving idiots as myself.