"Agnostic" plugin loading
(Somehow) manage to do exactly what I wanted from the start and now I no longer have to write each plugin into the body of App.ts. Hoping it won't break or shoot me in the foot or whatever. Still have to manually enter any new plugin into the ''/internal_modules/index.ts' though, which means the code is to be recompiled anyway.

Change "internal modules" into "plugins" because human language is hard for me, apparently.

Also change the command "recognition" a bit, so now it only checks (and reports to the log) recognized commands and not everything that starts with '/'.

Speaking of logs, yet to do that properly. Or at all.